+++ controllers/contacts.js
@@ -1,14 +1,7 @@
-const Joi = require("joi");
-
 const contactsService = require("../models/contacts");
 const { HttpErr } = require("../helpers/HttpErr");
 const { tryCatchWrapper } = require('../decorators');
 
-const contactAddSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
 const listContacts = async (req, res) => {
   const result = await contactsService.listContacts();
@@ -32,12 +25,6 @@ const getContactById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-  const { error } = contactAddSchema.validate(req.body);
-  // console.log(error);
-  if (error) {
-    throw HttpErr(400, error.message);
-  }
-  // console.log(req.body)
   const result = await contactsService.addContact(req.body);
   res.status(201).json(result);
 };
@@ -52,10 +39,6 @@ const removeContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
-  const { error } = contactAddSchema.validate(req.body);
-  if (error) {
-    throw HttpErr(400, error.message);
-  }
   const { contactId } = req.params;
   const result = await contactsService.updateContact(contactId, req.body);
   if (!result) {
