+++ decorators/index.js
@@ -1,5 +1,7 @@
 const tryCatchWrapper = require('./tryCatchlWrapper');
+const validateBody = require('./validateBody');
 
 module.exports = {
   tryCatchWrapper,
+  validateBody,
 };
