+++ routes/api/contacts.js
@@ -1,6 +1,8 @@
 const express = require('express');
 
 const contactsController = require("../../controllers/contacts");
+const schema = require('../../schemas/schema');
+const validateBody = require('../../decorators');
 
 const router = express.Router();
 
@@ -9,10 +11,10 @@ router.get('/', contactsController.listContacts);
 
 router.get('/:contactId', contactsController.getContactById);
 
-router.post('/', contactsController.addContact);
+router.post('/', validateBody(schema.contactAddSchema), contactsController.addContact);
 
 router.delete('/:contactId', contactsController.removeContact);
 
-router.put('/:contactId', contactsController.updateContact);
+router.put("/:contactId", validateBody(schema.contactAddSchema), contactsController.updateContact);
 
 module.exports = router
